Move fixtures to top level
